Split crumbs.js into pure crumb-parts and DOM crumb-renderer
The original `admin/app/views/browse/crumbs.js` mixed the Crumb stack
value type (parse/emit, label-resolution reads) with the DOM pillbar
renderer and async hydration. The split puts:

- `crumb-parts.js` — value type + cache-driven label readers, pure.
  Only dependencies are tunein-cache.js (label reads) and tunein-url.js
  (lcodeLabel for the language catalogue). No DOM, no api.js, no fetch.
- `crumb-renderer.js` — the renamed renderer half: pill-bar builder,
  filter badges, trail patching, hydrateCrumbLabels / hydrateFilterLabel
  / resolveLabelAndApply. Imports the pure half from `./crumb-parts.js`.

`outline-render.js` drops its local-copy `stringifyCrumbs` (the
cycle-avoidance workaround) and imports it cleanly from
`./crumb-parts.js` — the cycle is gone because crumb-parts.js has no
dependency on outline-render.js.

Tests follow the same seam: `test_crumb_parts.js` exercises the pure
half without the xmldom-backed dom-shim; `test_crumb_renderer.js`
covers the DOM + hydration surface. Coverage is preserved (27 + 39 = 66
assertions, matching the deleted `test_browse_crumbs.js`).
