+++ admin/app/views/browse/crumb-renderer.js
@@ -1,17 +1,19 @@
-// crumbs — the URL crumb stack as a self-contained module. Owns the
-// `parts` value type (the canonical drill-state shape passed around the
-// browser), the crumb token parser/emitter pair (parseCrumbs /
-// stringifyCrumbs, crumbTokenFor / partsFromCrumb), the pill-bar
-// renderer with its inline trail + filter badges (renderPillBar,
-// renderCrumbTrail, renderFilterBadges, renderFilterBadge), and the
-// async label hydration that upgrades raw-token renderings to
-// human-readable names once Describe/Browse resolves (hydrateCrumbLabels,
-// hydrateFilterLabel, patchTrailCrumb, patchBackAria,
-// resolveLabelAndApply).
+// crumb-renderer — DOM + async hydration half of the Crumb stack
+// module. The pure value type, parsing/emission, and label-resolution
+// reads live in `./crumb-parts.js`; this file builds on those to
+// render the pill bar (Back chevron + inline trail + filter badges)
+// and to upgrade raw-token renderings to human-readable names once
+// Describe / Browse resolves.
 //
-// The companion `views/browse.js` mounts the view; this module owns
-// everything below the mount boundary that deals with crumb-state
-// arithmetic.
+// What lives here:
+//   - DOM builders: renderPillBar, renderCrumbTrail, renderFilterBadges,
+//     renderFilterBadge (legacy single-filter shim), buildFilterBadge,
+//     makeSeparator.
+//   - Async label hydration: hydrateCrumbLabels, hydrateFilterLabel,
+//     resolveLabelAndApply, patchTrailCrumb, patchBackAria.
+//   - Small infrastructure: DESCRIBE_CONCURRENCY (concurrency cap for
+//     the cold-load label-fill wave), cssEscape (defensive attribute
+//     escape — CSS.escape isn't available under xmldom in tests).
 
 import { tuneinBrowse, tuneinDescribe } from '../../api.js';
 import { icon } from '../../icons.js';
@@ -19,209 +21,24 @@ import { lcodeLabel } from '../../tunein-url.js';
 import { cache, TTL_LABEL } from '../../tunein-cache.js';
 
 import { drillHashFor } from './outline-render.js';
-
-// Maximum depth of the URL crumb stack (`from=...`). Deeper than the
-// deepest observed location-tree depth (5); see issue #74 notes. Any
-// crumbs beyond this are dropped from the head of the stack.
-export const MAX_CRUMBS = 8;
+import {
+  TAB_LABEL_BY_TOKEN,
+  backAriaLabel,
+  backHrefFor,
+  crumbTokenFor,
+  filtersOf,
+  initialCrumbLabel,
+  initialFilterLabel,
+  initialHeaderTitle,
+  partsFromCrumb,
+  setFilters,
+} from './crumb-parts.js';
 
 // Maximum number of `Describe.ashx?id=<X>` calls in flight at once
 // during cold-load label-fill. Matches MAX_CRUMBS so the worst case
 // (a fully-deep shared URL with no cached labels) saturates in one
 // wave.
 const DESCRIBE_CONCURRENCY = 5;
-
-// Trail label override for the entry-point tab tokens. The first
-// crumb on any rooted drill is the tab — the spec wants its label
-// to read as the tab name (`Genre` / `Location` / `Language`), not
-// the cached API title ("Music" / "By Location" / "By Language").
-// Matched against the bare anchor portion of a crumb token; tokens
-// with an `:lXXX` filter suffix fall through to the cache /
-// catalogue path so e.g. `lang:l109` still resolves to "German".
-const TAB_LABEL_BY_TOKEN = Object.freeze({
-  music: 'Genre',
-  r0:    'Location',
-  lang:  'Language',
-});
-
-// ---- value type: parts + crumb token --------------------------------
-
-// Parse the comma-separated `from=` query parameter into an array of
-// crumb tokens. Empty / missing input returns []. Trims, drops empty
-// segments, caps to MAX_CRUMBS by keeping the tail (the most recent
-// crumbs — the ones closest to the current node).
-export function parseCrumbs(raw) {
-  if (typeof raw !== 'string' || raw === '') return [];
-  const segs = raw.split(',').map((s) => s.trim()).filter(Boolean);
-  return segs.length > MAX_CRUMBS ? segs.slice(segs.length - MAX_CRUMBS) : segs;
-}
-
-// Inverse of parseCrumbs: stringify the crumb array into a `from=`
-// value. Returns '' for an empty array so callers can skip emitting
-// the parameter entirely.
-export function stringifyCrumbs(crumbs) {
-  if (!Array.isArray(crumbs) || crumbs.length === 0) return '';
-  return crumbs.join(',');
-}
-
-// Helpers for the `filters` array shape. Always returns a fresh array
-// of trimmed non-empty strings. Tolerates legacy callers still passing
-// `parts.filter` as a single string (or a pre-joined comma list).
-export function filtersOf(parts) {
-  if (!parts) return [];
-  if (Array.isArray(parts.filters)) {
-    return parts.filters.filter((s) => typeof s === 'string' && s !== '');
-  }
-  if (typeof parts.filter === 'string' && parts.filter !== '') {
-    return parts.filter.split(',').map((s) => s.trim()).filter(Boolean);
-  }
-  return [];
-}
-
-// Stash a `filters: string[]` (and back-compat `filter: string`) on a
-// parts object. Empty arrays drop both fields so the URL composer emits
-// the bare drill (no `filter=` param).
-export function setFilters(parts, list) {
-  const cleaned = (Array.isArray(list) ? list : []).filter((s) => typeof s === 'string' && s !== '');
-  if (cleaned.length > 0) {
-    parts.filters = cleaned;
-    parts.filter  = cleaned.join(',');
-  } else {
-    delete parts.filters;
-    delete parts.filter;
-  }
-  return parts;
-}
-
-// Crumb token for a drill-parts object. The token is the value the
-// user would land on if they clicked Back to this level: prefer `id`,
-// fall back to `c`. When filters accompany the anchor we encode them
-// as `<anchor>:<f1>+<f2>+…` so two drills that share the same anchor
-// but differ in their filters (e.g. the language tree, where every
-// level emits `c=lang` or `c=music` with a different `filter=l<NNN>`)
-// produce distinct, navigable crumbs (#89, #106). pivots / offsets
-// remain refinements that do not become crumbs.
-// Returns null when neither anchor is set (root view).
-export function crumbTokenFor(parts) {
-  if (!parts) return null;
-  let anchor = '';
-  if (typeof parts.id === 'string' && parts.id !== '') anchor = parts.id;
-  else if (typeof parts.c === 'string' && parts.c !== '') anchor = parts.c;
-  if (!anchor) return null;
-  const filters = filtersOf(parts);
-  if (filters.length > 0) {
-    return `${anchor}:${filters.join('+')}`;
-  }
-  return anchor;
-}
-
-// Inverse of crumbTokenFor: turn a crumb token back into the drill
-// parts the user should land on. Heuristic:
-//   - bare tokens matching /^[a-z]\d+$/ are guide_ids (`id=` anchor)
-//   - bare tokens otherwise are category short names (`c=`)
-//   - tokens with a `:f1+f2+…` suffix attach N filters to whichever
-//     anchor form applies (so `lang:l109` → `{c:'lang', filters:['l109']}`,
-//     `r101821:g26+l170` → `{id:'r101821', filters:['g26','l170']}`)
-export function partsFromCrumb(token) {
-  if (typeof token !== 'string' || token === '') return null;
-  const colonIdx = token.indexOf(':');
-  let anchor = token;
-  let filterRaw = '';
-  if (colonIdx >= 0) {
-    anchor = token.slice(0, colonIdx);
-    filterRaw = token.slice(colonIdx + 1);
-  }
-  if (!anchor) return null;
-  const parts = /^[a-z]\d+$/.test(anchor) ? { id: anchor } : { c: anchor };
-  if (filterRaw) {
-    const list = filterRaw.split('+').map((s) => s.trim()).filter(Boolean);
-    setFilters(parts, list);
-  }
-  return parts;
-}
-
-// A drill URL can carry any of {id, c, filter, pivot, offset}. The
-// presence of `id` or `c` is the load-bearing signal; the others
-// modify the drill. The URL's `filter=` may be comma-separated values
-// (multi-filter wire shape, #106); the resulting parts object carries
-// `filters: string[]` plus back-compat `filter: string`. Returns null
-// when neither anchor is set (root view).
-export function pickDrillParts(query) {
-  if (!query || (typeof query.id !== 'string' && typeof query.c !== 'string')) {
-    return null;
-  }
-  const out = {};
-  for (const key of ['id', 'c', 'pivot', 'offset']) {
-    if (typeof query[key] === 'string' && query[key] !== '') out[key] = query[key];
-  }
-  if (typeof query.filter === 'string' && query.filter !== '') {
-    const list = query.filter.split(',').map((s) => s.trim()).filter(Boolean);
-    setFilters(out, list);
-  }
-  return out;
-}
-
-// Display label for the drill crumb — a compact form of the parts.
-// `c=music&filter=l216` reads more usefully than just `music`. When
-// the filter is an lcode (`l<NNN>`) and we have a cached language
-// name for it, append the human form so end users get an at-a-glance
-// translation of the otherwise opaque numeric filter (#90). With
-// multi-filter drills (#106) the comma-joined wire value is shown;
-// a single-lcode filter still resolves to its language name.
-export function crumbLabelFor(parts) {
-  const segs = [];
-  const filters = filtersOf(parts);
-  const filterStr = filters.join(',');
-  if (parts.id) segs.push(parts.id);
-  if (parts.c)  segs.push(`c=${parts.c}`);
-  if (filterStr) segs.push(`filter=${filterStr}`);
-  if (parts.pivot)  segs.push(`pivot=${parts.pivot}`);
-  if (parts.offset) segs.push(`offset=${parts.offset}`);
-  let label = segs.join(' · ');
-  if (filters.length === 1 && /^l\d+$/.test(filters[0])) {
-    const name = lcodeLabel(filters[0]);
-    if (name) label += ` (${name})`;
-  }
-  return label;
-}
-
-// Compose the Back-button href. Pops the rightmost crumb and navigates
-// to it; the popped crumb keeps the remaining stack as its own `from=`.
-// An empty stack lands at #/browse (the root tabs view).
-export function backHrefFor(stack) {
-  if (!Array.isArray(stack) || stack.length === 0) return '#/browse';
-  const head = stack.slice(0, -1);
-  const tail = stack[stack.length - 1];
-  const parts = partsFromCrumb(tail);
-  if (!parts) return '#/browse';
-  // Always pass the trimmed crumb stack explicitly so we don't pick up
-  // a stale value of _childCrumbs (set later, after this Back-link is
-  // built — but we want backHref deterministic regardless of order).
-  return drillHashFor(parts, head);
-}
-
-// Initial page-header title before the API response arrives. Prefer
-// the cached label for this node's token; fall back to crumbLabelFor
-// so the user sees *something* during the load.
-//
-// Filter-bearing tokens (`<anchor>:<filter>`) try the bare anchor too:
-// the page IS the anchor node (just filtered), so a cached title for
-// `r101821` should surface as the h1 / current-crumb text for the
-// `r101821:g26` drill too. The filter surface is the badge, not the
-// title.
-export function initialHeaderTitle(parts, token) {
-  if (token) {
-    const cached = cache.get(`tunein.label.${token}`);
-    if (typeof cached === 'string' && cached !== '') return cached;
-    const colon = token.indexOf(':');
-    if (colon > 0) {
-      const bare = token.slice(0, colon);
-      const cachedBare = cache.get(`tunein.label.${bare}`);
-      if (typeof cachedBare === 'string' && cachedBare !== '') return cachedBare;
-    }
-  }
-  return crumbLabelFor(parts);
-}
 
 // ---- pill bar + inline trail (top of every drill body) -------------
 //
@@ -316,7 +133,7 @@ export function renderFilterBadges(parts, stack) {
 // Build one badge node for `filterToken`, given the full filter list
 // and current parts / stack so the × close anchor can drop only this
 // filter while preserving the rest.
-function buildFilterBadge(filterToken, parts, filters, stack) {
+export function buildFilterBadge(filterToken, parts, filters, stack) {
   const badge = document.createElement('span');
   badge.className = 'browse-bar__filter';
   badge.dataset.filterToken = filterToken;
@@ -361,20 +178,6 @@ function buildFilterBadge(filterToken, parts, filters, stack) {
 export function renderFilterBadge(parts, stack) {
   const badges = renderFilterBadges(parts, stack);
   return badges.length > 0 ? badges[0] : null;
-}
-
-// Pick the best already-known label for a filter token without any
-// network work. Mirrors the trail's `initialCrumbLabel` ladder for the
-// filter-only case: lcode catalogue for `lXXX`, then the cache, then
-// the raw token as a last resort.
-function initialFilterLabel(filterToken) {
-  if (/^l\d+$/.test(filterToken)) {
-    const name = lcodeLabel(filterToken);
-    if (name) return name;
-  }
-  const cached = cache.get(`tunein.label.${filterToken}`);
-  if (typeof cached === 'string' && cached !== '') return cached;
-  return filterToken;
 }
 
 // One-shot label fetch for a filter token whose initial render is the
@@ -512,56 +315,6 @@ export function makeSeparator() {
   sep.setAttribute('aria-hidden', 'true');
   sep.textContent = '›'; // ›
   return sep;
-}
-
-// Pick the best already-known label for a crumb token without doing
-// any network work. Order of preference for ancestors:
-//   - tab-token override (only at stack[0] — the entry-point tab)
-//   - cached label under `tunein.label.<token>` (from a previous
-//     resolve)
-//   - in-memory lcode catalogue (free for `<anchor>:l<NNN>` tokens)
-//   - raw token fallback
-// The async hydrateCrumbLabels path upgrades any token still
-// rendered as raw once its Describe / Browse lookup completes.
-export function initialCrumbLabel(token, parts, isFirstInStack) {
-  // The first crumb in the stack is the entry-point tab — render it
-  // with the tab's display label (`Genre` / `Location` / `Language`)
-  // rather than the cached API title ("Music" / "By Location" / "By
-  // Language"). Filter-bearing tokens (`lang:lXXX`) intentionally
-  // fall through to the catalogue path so they read as the language
-  // name, not "Language".
-  if (isFirstInStack) {
-    const bareAnchor = parts && typeof parts.id === 'string'
-      ? parts.id
-      : (parts && typeof parts.c === 'string' ? parts.c : '');
-    const isFilterBearing = parts && typeof parts.filter === 'string' && parts.filter !== '';
-    if (!isFilterBearing && bareAnchor && TAB_LABEL_BY_TOKEN[bareAnchor]) {
-      return TAB_LABEL_BY_TOKEN[bareAnchor];
-    }
-  }
-  const cached = cache.get(`tunein.label.${token}`);
-  if (typeof cached === 'string' && cached !== '') return cached;
-  if (parts && typeof parts.filter === 'string' && /^l\d+$/.test(parts.filter)) {
-    const name = lcodeLabel(parts.filter);
-    if (name) return name;
-  }
-  return token;
-}
-
-// Compose the Back chevron's aria-label from the popped destination.
-// Empty stack → root tabs view ("Back to Browse"). Otherwise the
-// destination is the last crumb in the stack; prefer the resolved
-// label (cache / catalogue) and fall back to the raw token.
-export function backAriaLabel(stack) {
-  if (!Array.isArray(stack) || stack.length === 0) return 'Back to Browse';
-  const destToken = stack[stack.length - 1];
-  if (!destToken) return 'Back to Browse';
-  const destParts = partsFromCrumb(destToken);
-  // The last crumb in the stack is the trail's tail-before-current —
-  // it's an ancestor, not the entry-point tab, so isFirstInStack is
-  // only relevant when stack length is 1.
-  const label = initialCrumbLabel(destToken, destParts, stack.length === 1);
-  return `Back to ${label}`;
 }
 
 // ---- hydration: resolve raw tokens to human-readable labels --------
